requirements.txt
python-dotenv
streamlit
langchain
langchain-openai
faiss-cpu
pypdf
tiktoken
sentence-transformers
reportlab